pyproject.toml
[project]
name = "qcml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jax>=0.8.1",
    "jaxopt>=0.8.5",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "optax>=0.2.6",
    "pyyaml>=6.0.3",
]
